SDL2 -> SDL3 (#68), kinda broken, but builds Uint64 ticks (except over network, TODO) [still needs sanity checks], no audio, transparency broken on some textures, cmake change is an educated guess
TODOs added for SDL3_mixer release, should be releasing in a few weeks?

+++ src/fpsgame/fragmessages.cpp
@@ -51,7 +51,7 @@ namespace game {
                 break;
             }
 
-            int alpha = 255 - max(0, lastmillis-m.fragtime-showmillis);
+            int alpha = 255 - max((Uint64)0, lastmillis - m.fragtime - showmillis);
 
             vec2 drawposcenter = vec2(0, (fragmessages.length()-1-i)*stackdir*stepsize).add(origin);
