+++ src/fpsgame/scoreboard.cpp
@@ -204,7 +204,7 @@ namespace game
 			if(intermission) g.text("intermission", COL_WHITE);
 			else
 			{
-				int secs = max(maplimit - lastmillis + 999, 0) / 1000, mins = secs / 60;
+				int secs = max(maplimit - lastmillis + 999, (Uint64)0) / 1000, mins = secs / 60;
 				secs %= 60;
 				g.pushlist();
 				g.strut(mins >= 10 ? 4.5f : 3.5f);
